test: convert to vite
